snap scroll: add prevPrevDeck
This makes the scroll-up animation smoother by showing the preview of the previous card. However, I think this my not be tenable from a performance POV – I noticed the expansion animation dropped frames more frequently with this change. Want to see if other's can repro and talk to Ben about potential solutions